+++ packages/assets-controllers/src/CurrencyRateController.test.ts
@@ -284,7 +284,7 @@ describe('CurrencyRateController', () => {
     expect(controller.state.currencyRates).toStrictEqual({
       ETH: {
         conversionDate: 10,
-        conversionRate: 4149.76,
+        conversionRate: 4149.764437073889,
         usdConversionRate: null,
       },
     });
@@ -298,7 +298,7 @@ describe('CurrencyRateController', () => {
     expect(controller.state.currencyRates).toStrictEqual({
       ETH: {
         conversionDate: 20,
-        conversionRate: 4149.76,
+        conversionRate: 4149.764437073889,
         usdConversionRate: null,
       },
     });
@@ -417,8 +417,8 @@ describe('CurrencyRateController', () => {
     expect(controller.state.currencyRates).toStrictEqual({
       ETH: {
         conversionDate: getStubbedDate() / 1000,
-        conversionRate: 4149.76,
-        usdConversionRate: 0.01,
+        conversionRate: 4149.764437073889,
+        usdConversionRate: 0.009009009009009009,
       },
     });
 
@@ -469,7 +469,7 @@ describe('CurrencyRateController', () => {
       },
       SepoliaETH: {
         conversionDate: getStubbedDate() / 1000,
-        conversionRate: 4149.76,
+        conversionRate: 4149.764437073889,
         usdConversionRate: 1000,
       },
     });
@@ -538,13 +538,13 @@ describe('CurrencyRateController', () => {
       currencyRates: {
         ETH: {
           conversionDate: getStubbedDate() / 1000,
-          conversionRate: 4149.76,
-          usdConversionRate: 181.82,
+          conversionRate: 4149.764437073889,
+          usdConversionRate: 181.81818181818184,
         },
         BTC: {
           conversionDate: getStubbedDate() / 1000,
-          conversionRate: 9636.65,
-          usdConversionRate: 454.55,
+          conversionRate: 9636.6518,
+          usdConversionRate: 454.5454545454545,
         },
       },
     });
@@ -617,7 +617,7 @@ describe('CurrencyRateController', () => {
       currencyRates: {
         ETH: {
           conversionDate: getStubbedDate() / 1000,
-          conversionRate: 2000.42,
+          conversionRate: 2000.4199999999998,
           usdConversionRate: null,
         },
       },
@@ -1139,7 +1139,6 @@ describe('CurrencyRateController', () => {
       jest
         .spyOn(tokenPricesService, 'fetchTokenPrices')
         .mockImplementation(async ({ assets }) => {
-          // eslint-disable-next-line jest/no-conditional-in-test
           if (assets.some((asset) => asset.chainId === '0x1')) {
             return [
               {
@@ -1168,7 +1167,7 @@ describe('CurrencyRateController', () => {
               },
             ];
           }
-          // eslint-disable-next-line jest/no-conditional-in-test
+
           if (assets.some((asset) => asset.chainId === '0x89')) {
             return [
               {
@@ -1262,11 +1261,9 @@ describe('CurrencyRateController', () => {
       const fetchTokenPricesSpy = jest
         .spyOn(tokenPricesService, 'fetchTokenPrices')
         .mockImplementation(async ({ assets }) => {
-          // eslint-disable-next-line jest/no-conditional-in-test
           if (
             assets.some(
               (asset) =>
-                // eslint-disable-next-line jest/no-conditional-in-test
                 asset.chainId === '0x1' || asset.chainId === '0xaa36a7',
             )
           ) {
@@ -1358,7 +1355,6 @@ describe('CurrencyRateController', () => {
       jest
         .spyOn(tokenPricesService, 'fetchTokenPrices')
         .mockImplementation(async ({ assets }) => {
-          // eslint-disable-next-line jest/no-conditional-in-test
           if (assets.some((asset) => asset.chainId === '0x1')) {
             // ETH succeeds
             return [
@@ -1498,7 +1494,6 @@ describe('CurrencyRateController', () => {
       const fetchTokenPricesSpy = jest
         .spyOn(tokenPricesService, 'fetchTokenPrices')
         .mockImplementation(async ({ assets }) => {
-          // eslint-disable-next-line jest/no-conditional-in-test
           if (assets.some((asset) => asset.chainId === '0x1')) {
             return [
               {