+++ packages/network-controller/tests/NetworkController.test.ts
@@ -33,6 +33,7 @@ import {
   buildUpdateNetworkCustomRpcEndpointFields,
   INFURA_NETWORKS,
   TESTNET,
+  withController,
 } from './helpers';
 import type { RootMessenger } from './helpers';
 import { FakeBlockTracker } from '../../../tests/fake-block-tracker';
@@ -50,8 +51,6 @@ import type {
   NetworkClientId,
   NetworkConfiguration,
   NetworkControllerEvents,
-  NetworkControllerMessenger,
-  NetworkControllerOptions,
   NetworkControllerStateChangeEvent,
   NetworkState,
 } from '../src/NetworkController';
@@ -508,6 +507,21 @@ describe('NetworkController', () => {
                   },
                 ],
               },
+              "0x279f": Object {
+                "blockExplorerUrls": Array [],
+                "chainId": "0x279f",
+                "defaultRpcEndpointIndex": 0,
+                "name": "Monad Testnet",
+                "nativeCurrency": "MON",
+                "rpcEndpoints": Array [
+                  Object {
+                    "failoverUrls": Array [],
+                    "networkClientId": "monad-testnet",
+                    "type": "infura",
+                    "url": "https://monad-testnet.infura.io/v3/{infuraProjectId}",
+                  },
+                ],
+              },
               "0x38": Object {
                 "blockExplorerUrls": Array [],
                 "chainId": "0x38",
@@ -692,6 +706,21 @@ describe('NetworkController', () => {
                       "networkClientId": "base-mainnet",
                       "type": "infura",
                       "url": "https://base-mainnet.infura.io/v3/{infuraProjectId}",
+                    },
+                  ],
+                },
+                "0x279f": Object {
+                  "blockExplorerUrls": Array [],
+                  "chainId": "0x279f",
+                  "defaultRpcEndpointIndex": 0,
+                  "name": "Monad Testnet",
+                  "nativeCurrency": "MON",
+                  "rpcEndpoints": Array [
+                    Object {
+                      "failoverUrls": Array [],
+                      "networkClientId": "monad-testnet",
+                      "type": "infura",
+                      "url": "https://monad-testnet.infura.io/v3/{infuraProjectId}",
                     },
                   ],
                 },
@@ -1995,6 +2024,21 @@ describe('NetworkController', () => {
                   chainId: '0x1',
                   ticker: 'ETH',
                   network: InfuraNetworkType.mainnet,
+                },
+                provider: expect.anything(),
+                destroy: expect.any(Function),
+                enableRpcFailover: expect.any(Function),
+                disableRpcFailover: expect.any(Function),
+              },
+              'monad-testnet': {
+                blockTracker: expect.anything(),
+                configuration: {
+                  type: NetworkClientType.Infura,
+                  failoverRpcUrls: [],
+                  infuraProjectId,
+                  chainId: '0x279f',
+                  ticker: 'MON',
+                  network: InfuraNetworkType['monad-testnet'],
                 },
                 provider: expect.anything(),
                 destroy: expect.any(Function),
@@ -4637,6 +4681,7 @@ describe('NetworkController', () => {
               expect(createAutoManagedNetworkClientSpy).toHaveBeenNthCalledWith(
                 2,
                 {
+                  networkClientId: infuraNetworkType,
                   networkClientConfiguration: {
                     infuraProjectId,
                     failoverRpcUrls: ['https://first.failover.endpoint'],
@@ -4654,6 +4699,7 @@ describe('NetworkController', () => {
               expect(createAutoManagedNetworkClientSpy).toHaveBeenNthCalledWith(
                 3,
                 {
+                  networkClientId: 'BBBB-BBBB-BBBB-BBBB',
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://second.failover.endpoint'],
@@ -4670,6 +4716,7 @@ describe('NetworkController', () => {
               expect(createAutoManagedNetworkClientSpy).toHaveBeenNthCalledWith(
                 4,
                 {
+                  networkClientId: 'CCCC-CCCC-CCCC-CCCC',
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://third.failover.endpoint'],
@@ -6047,6 +6094,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(3, {
+                  networkClientId: infuraNetworkType,
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://failover.endpoint'],
@@ -6278,6 +6326,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(3, {
+                  networkClientId: 'AAAA-AAAA-AAAA-AAAA',
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://first.failover.endpoint'],
@@ -6293,6 +6342,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(4, {
+                  networkClientId: 'BBBB-BBBB-BBBB-BBBB',
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://second.failover.endpoint'],
@@ -7265,6 +7315,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(3, {
+                  networkClientId: 'BBBB-BBBB-BBBB-BBBB',
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://failover.endpoint'],
@@ -8135,6 +8186,7 @@ describe('NetworkController', () => {
               expect(createAutoManagedNetworkClientSpy).toHaveBeenNthCalledWith(
                 3,
                 {
+                  networkClientId: 'BBBB-BBBB-BBBB-BBBB',
                   networkClientConfiguration: {
                     chainId: '0x1337',
                     failoverRpcUrls: ['https://first.failover.endpoint'],
@@ -8151,6 +8203,7 @@ describe('NetworkController', () => {
               expect(createAutoManagedNetworkClientSpy).toHaveBeenNthCalledWith(
                 4,
                 {
+                  networkClientId: 'CCCC-CCCC-CCCC-CCCC',
                   networkClientConfiguration: {
                     chainId: '0x1337',
                     failoverRpcUrls: ['https://second.failover.endpoint'],
@@ -9136,6 +9189,7 @@ describe('NetworkController', () => {
               });
 
               expect(createAutoManagedNetworkClientSpy).toHaveBeenCalledWith({
+                networkClientId: 'BBBB-BBBB-BBBB-BBBB',
                 networkClientConfiguration: {
                   chainId: '0x1337',
                   failoverRpcUrls: ['https://failover.endpoint'],
@@ -10292,6 +10346,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(4, {
+                  networkClientId: 'CCCC-CCCC-CCCC-CCCC',
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://first.failover.endpoint'],
@@ -10307,6 +10362,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(5, {
+                  networkClientId: 'DDDD-DDDD-DDDD-DDDD',
                   networkClientConfiguration: {
                     chainId: infuraChainId,
                     failoverRpcUrls: ['https://second.failover.endpoint'],
@@ -11008,6 +11064,7 @@ describe('NetworkController', () => {
                 );
 
                 expect(createAutoManagedNetworkClientSpy).toHaveBeenCalledWith({
+                  networkClientId: 'CCCC-CCCC-CCCC-CCCC',
                   networkClientConfiguration: {
                     chainId: '0x1337',
                     failoverRpcUrls: ['https://first.failover.endpoint'],
@@ -11021,6 +11078,7 @@ describe('NetworkController', () => {
                   isRpcFailoverEnabled: true,
                 });
                 expect(createAutoManagedNetworkClientSpy).toHaveBeenCalledWith({
+                  networkClientId: 'DDDD-DDDD-DDDD-DDDD',
                   networkClientConfiguration: {
                     chainId: '0x1337',
                     failoverRpcUrls: ['https://second.failover.endpoint'],
@@ -11737,6 +11795,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(6, {
+                  networkClientId: 'CCCC-CCCC-CCCC-CCCC',
                   networkClientConfiguration: {
                     chainId: anotherInfuraChainId,
                     failoverRpcUrls: ['https://first.failover.endpoint'],
@@ -11752,6 +11811,7 @@ describe('NetworkController', () => {
                 expect(
                   createAutoManagedNetworkClientSpy,
                 ).toHaveBeenNthCalledWith(7, {
+                  networkClientId: 'DDDD-DDDD-DDDD-DDDD',
                   networkClientConfiguration: {
                     chainId: anotherInfuraChainId,
                     failoverRpcUrls: ['https://second.failover.endpoint'],
@@ -12434,6 +12494,7 @@ describe('NetworkController', () => {
             expect(createAutoManagedNetworkClientSpy).toHaveBeenNthCalledWith(
               4,
               {
+                networkClientId: 'CCCC-CCCC-CCCC-CCCC',
                 networkClientConfiguration: {
                   chainId: '0x2448',
                   failoverRpcUrls: ['https://first.failover.endpoint'],
@@ -12450,6 +12511,7 @@ describe('NetworkController', () => {
             expect(createAutoManagedNetworkClientSpy).toHaveBeenNthCalledWith(
               5,
               {
+                networkClientId: 'DDDD-DDDD-DDDD-DDDD',
                 networkClientConfiguration: {
                   chainId: '0x2448',
                   failoverRpcUrls: ['https://second.failover.endpoint'],
@@ -14793,6 +14855,21 @@ describe('NetworkController', () => {
                   },
                 ],
               },
+              "0x279f": Object {
+                "blockExplorerUrls": Array [],
+                "chainId": "0x279f",
+                "defaultRpcEndpointIndex": 0,
+                "name": "Monad Testnet",
+                "nativeCurrency": "MON",
+                "rpcEndpoints": Array [
+                  Object {
+                    "failoverUrls": Array [],
+                    "networkClientId": "monad-testnet",
+                    "type": "infura",
+                    "url": "https://monad-testnet.infura.io/v3/{infuraProjectId}",
+                  },
+                ],
+              },
               "0x38": Object {
                 "blockExplorerUrls": Array [],
                 "chainId": "0x38",
@@ -14962,6 +15039,21 @@ describe('NetworkController', () => {
                   },
                 ],
               },
+              "0x279f": Object {
+                "blockExplorerUrls": Array [],
+                "chainId": "0x279f",
+                "defaultRpcEndpointIndex": 0,
+                "name": "Monad Testnet",
+                "nativeCurrency": "MON",
+                "rpcEndpoints": Array [
+                  Object {
+                    "failoverUrls": Array [],
+                    "networkClientId": "monad-testnet",
+                    "type": "infura",
+                    "url": "https://monad-testnet.infura.io/v3/{infuraProjectId}",
+                  },
+                ],
+              },
               "0x38": Object {
                 "blockExplorerUrls": Array [],
                 "chainId": "0x38",
@@ -15128,6 +15220,21 @@ describe('NetworkController', () => {
                     "networkClientId": "base-mainnet",
                     "type": "infura",
                     "url": "https://base-mainnet.infura.io/v3/{infuraProjectId}",
+                  },
+                ],
+              },
+              "0x279f": Object {
+                "blockExplorerUrls": Array [],
+                "chainId": "0x279f",
+                "defaultRpcEndpointIndex": 0,
+                "name": "Monad Testnet",
+                "nativeCurrency": "MON",
+                "rpcEndpoints": Array [
+                  Object {
+                    "failoverUrls": Array [],
+                    "networkClientId": "monad-testnet",
+                    "type": "infura",
+                    "url": "https://monad-testnet.infura.io/v3/{infuraProjectId}",
                   },
                 ],
               },
@@ -16569,53 +16676,6 @@ function lookupNetworkTests({
       });
     }
   });
-}
-
-type WithControllerCallback<ReturnValue> = ({
-  controller,
-}: {
-  controller: NetworkController;
-  messenger: RootMessenger;
-  networkControllerMessenger: NetworkControllerMessenger;
-}) => Promise<ReturnValue> | ReturnValue;
-
-type WithControllerOptions = Partial<NetworkControllerOptions>;
-
-type WithControllerArgs<ReturnValue> =
-  | [WithControllerCallback<ReturnValue>]
-  | [WithControllerOptions, WithControllerCallback<ReturnValue>];
-
-/**
- * Builds a controller based on the given options, and calls the given function
- * with that controller.
- *
- * @param args - Either a function, or an options bag + a function. The options
- * bag is equivalent to the options that NetworkController takes (although
- * `messenger` and `infuraProjectId` are  filled in if not given); the function
- * will be called with the built controller.
- * @returns Whatever the callback returns.
- */
-async function withController<ReturnValue>(
-  ...args: WithControllerArgs<ReturnValue>
-): Promise<ReturnValue> {
-  const [{ ...rest }, fn] = args.length === 2 ? args : [{}, args[0]];
-  const messenger = buildRootMessenger();
-  const networkControllerMessenger = buildNetworkControllerMessenger(messenger);
-  const controller = new NetworkController({
-    messenger: networkControllerMessenger,
-    infuraProjectId: 'infura-project-id',
-    getRpcServiceOptions: () => ({
-      fetch,
-      btoa,
-    }),
-    ...rest,
-  });
-  try {
-    return await fn({ controller, messenger, networkControllerMessenger });
-  } finally {
-    const { blockTracker } = controller.getProviderAndBlockTracker();
-    await blockTracker?.destroy();
-  }
 }
 
 /**