+++ packages/subscription-controller/src/types.ts
@@ -133,6 +133,12 @@ export type StartSubscriptionRequest = {
   recurringInterval: RecurringInterval;
   successUrl?: string;
   useTestClock?: boolean;
+
+  /**
+   * The optional ID of the reward subscription to be opt in along with the main `shield` subscription.
+   * This is required if user wants to opt in to the reward subscription during the `shield` subscription creation.
+   */
+  rewardSubscriptionId?: string;
 };
 
 export type StartSubscriptionResponse = {
@@ -153,11 +159,31 @@ export type StartCryptoSubscriptionRequest = {
   rawTransaction: Hex;
   isSponsored?: boolean;
   useTestClock?: boolean;
+  /**
+   * The optional ID of the reward subscription to be opt in along with the main `shield` subscription.
+   * This is required if user wants to opt in to the reward subscription during the `shield` subscription creation.
+   */
+  rewardSubscriptionId?: string;
 };
 
 export type StartCryptoSubscriptionResponse = {
   subscriptionId: string;
   status: SubscriptionStatus;
+};
+
+/**
+ * General response type for the subscription API requests
+ * which doesn't require any specific response data.
+ */
+export type SubscriptionApiGeneralResponse = {
+  /**
+   * Whether the request was successful.
+   */
+  success: boolean;
+  /**
+   * The message of the response.
+   */
+  message?: string;
 };
 
 export type AuthUtils = {
@@ -171,7 +197,10 @@ export type ProductPrice = {
   /** only usd for now */
   currency: Currency;
   trialPeriodDays: number;
+  /** min billing cycles for approval */
   minBillingCycles: number;
+  /** min billing cycles for account balance check */
+  minBillingCyclesForBalance: number;
 };
 
 export type ProductPricing = {
@@ -273,7 +302,6 @@ export type Cohort = {
 export type SubscriptionEligibility = {
   product: ProductType;
   canSubscribe: boolean;
-  minBalanceUSD: number;
   canViewEntryModal: boolean;
   modalType?: ModalType;
   cohorts: Cohort[];
@@ -348,6 +376,13 @@ export type ISubscriptionService = {
   assignUserToCohort(request: AssignCohortRequest): Promise<void>;
 
   /**
+   * Link rewards to a subscription.
+   */
+  linkRewards(
+    request: LinkRewardsRequest,
+  ): Promise<SubscriptionApiGeneralResponse>;
+
+  /**
    * Submit sponsorship intents to the Subscription Service backend.
    *
    * This is intended to be used together with the crypto subscription flow.
@@ -419,4 +454,14 @@ export type CachedLastSelectedPaymentMethod = {
   paymentTokenSymbol?: string;
   plan: RecurringInterval;
   useTestClock?: boolean;
+};
+
+/**
+ * Request object for linking rewards to a subscription.
+ */
+export type LinkRewardsRequest = {
+  /**
+   * The ID of the reward subscription to be linked to the subscription.
+   */
+  rewardSubscriptionId: string;
 };