+++ packages/transaction-controller/src/TransactionController.test.ts
@@ -1131,6 +1131,72 @@ describe('TransactionController', () => {
       expect(transactions[5].status).toBe(TransactionStatus.failed);
     });
 
+    it('fails required transactions of approved and signed transactions', async () => {
+      const mockTransactionMeta = {
+        from: ACCOUNT_MOCK,
+        txParams: {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_2_MOCK,
+        },
+      };
+      const mockedTransactions = [
+        {
+          id: '123',
+          history: [{ ...mockTransactionMeta, id: '123' }],
+          chainId: toHex(5),
+          status: TransactionStatus.approved,
+          requiredTransactionIds: ['222', '333'],
+          ...mockTransactionMeta,
+        },
+        {
+          id: '111',
+          history: [{ ...mockTransactionMeta, id: '111' }],
+          chainId: toHex(5),
+          status: TransactionStatus.signed,
+          ...mockTransactionMeta,
+        },
+        {
+          id: '222',
+          history: [{ ...mockTransactionMeta, id: '222' }],
+          chainId: toHex(1),
+          status: TransactionStatus.confirmed,
+          ...mockTransactionMeta,
+        },
+        {
+          id: '333',
+          history: [{ ...mockTransactionMeta, id: '333' }],
+          chainId: toHex(16),
+          status: TransactionStatus.submitted,
+          ...mockTransactionMeta,
+        },
+      ];
+
+      const mockedControllerState = {
+        transactions: mockedTransactions,
+        methodData: {},
+        lastFetchedBlockNumbers: {},
+      };
+
+      const { controller } = setupController({
+        options: {
+          // TODO: Replace `any` with type
+          // eslint-disable-next-line @typescript-eslint/no-explicit-any
+          state: mockedControllerState as any,
+        },
+      });
+
+      await flushPromises();
+
+      const { transactions } = controller.state;
+
+      expect(transactions.map((t) => [t.id, t.status])).toStrictEqual([
+        ['333', TransactionStatus.failed],
+        ['222', TransactionStatus.confirmed],
+        ['111', TransactionStatus.failed],
+        ['123', TransactionStatus.failed],
+      ]);
+    });
+
     it('removes unapproved transactions', async () => {
       const mockTransactionMeta = {
         from: ACCOUNT_MOCK,
@@ -8498,7 +8564,7 @@ describe('TransactionController', () => {
         expect(result).toStrictEqual([GAS_FEE_TOKEN_MOCK]);
       });
 
-      it('includes delegation address in request', async () => {
+      it('includes delegation address and isExternalSign in request', async () => {
         const { messenger } = setupController();
 
         getGasFeeTokensMock.mockResolvedValueOnce({
@@ -8520,6 +8586,7 @@ describe('TransactionController', () => {
           expect.objectContaining({
             transactionMeta: expect.objectContaining({
               delegationAddress: ACCOUNT_2_MOCK,
+              isExternalSign: true,
             }),
           }),
         );